# backend/storage/api-docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Explus API",
        "version": "1.0.0",
        "description": "API n\u1ed9i b\u1ed9 c\u1ee7a Explus \u2014 qu\u1ea3n l\u00fd c\u00f4ng vi\u1ec7c, d\u1ef1 \u00e1n, b\u00ecnh lu\u1eadn v\u00e0 th\u00f4ng b\u00e1o.\n\nX\u00e1c th\u1ef1c b\u1eb1ng **cookie phi\u00ean** c\u1ee7a Laravel Sanctum, kh\u00f4ng d\u00f9ng token trong header: l\u1ea5y cookie CSRF \u1edf `/sanctum/csrf-cookie`, \u0111\u0103ng nh\u1eadp \u1edf `/api/v1/auth/login`, r\u1ed3i x\u00e1c nh\u1eadn m\u00e3 OTP \u1edf `/api/v1/auth/two-factor-challenge`. X\u00e1c th\u1ef1c hai l\u1edbp l\u00e0 **b\u1eaft bu\u1ed9c v\u1edbi m\u1ecdi t\u00e0i kho\u1ea3n**.\n\nM\u1ecdi l\u1ed7i tr\u1ea3 v\u1ec1 c\u00f9ng m\u1ed9t d\u1ea1ng: `{message, code, errors}`. M\u1ecdi m\u1ed1c th\u1eddi gian tr\u1ea3 v\u1ec1 theo ISO 8601 k\u00e8m offset; khi g\u1eedi l\u00ean, chu\u1ed7i kh\u00f4ng k\u00e8m offset \u0111\u01b0\u1ee3c hi\u1ec3u l\u00e0 gi\u1edd Vi\u1ec7t Nam.\n\n\u0110\u1ecbnh danh l\u1ed9 ra ngo\u00e0i lu\u00f4n l\u00e0 uuid, kh\u00f4ng ph\u1ea3i id tu\u1ea7n t\u1ef1."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/comments/{comment}/attachments/{media}": {
            "delete": {
                "operationId": "commentAttachment.destroy",
                "tags": [
                    "CommentAttachment"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "description": "The comment UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "media",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/projects/{project}": {
            "get": {
                "operationId": "project.show",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`ProjectResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProjectResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "owner",
                                                        "department"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "project.update",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`ProjectResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/ProjectResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "owner",
                                                        "department"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "project.destroy",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/projects/{project}/members/{user}": {
            "delete": {
                "operationId": "projectMember.destroy",
                "tags": [
                    "ProjectMember"
                ],
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "description": "The project UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "description": "The user UUID",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ]
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/tasks/{task}": {
            "get": {
                "operationId": "task.show",
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`TaskResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/TaskResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "project",
                                                        "assignee",
                                                        "assigner",
                                                        "reviewer",
                                                        "labels"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "patch": {
                "operationId": "task.update",
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`TaskResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/TaskResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "project",
                                                        "assignee",
                                                        "assigner"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "task.destroy",
                "tags": [
                    "Task"
                ],
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "description": "The task UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/comments/{comment}": {
            "patch": {
                "operationId": "taskComment.update",
                "tags": [
                    "TaskComment"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "description": "The comment UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCommentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`TaskCommentResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/TaskCommentResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "author",
                                                        "mentions"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "taskComment.destroy",
                "tags": [
                    "TaskComment"
                ],
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "description": "The comment UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "L\u1ed9 uuid, kh\u00f4ng l\u1ed9 id tu\u1ea7n t\u1ef1 \u2014 xem README \"Quy \u01b0\u1edbc d\u1eef li\u1ec7u\"."
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "status": {
                        "type": "object",
                        "properties": {
                            "value": {
                                "type": "string"
                            },
                            "label": {
                                "type": "string",
                                "enum": [
                                    "\u0110ang l\u00ean k\u1ebf ho\u1ea1ch",
                                    "\u0110ang ch\u1ea1y",
                                    "T\u1ea1m d\u1eebng",
                                    "Ho\u00e0n th\u00e0nh",
                                    "\u0110\u00e3 hu\u1ef7"
                                ]
                            },
                            "is_open": {
                                "type": "boolean"
                            }
                        },
                        "required": [
                            "value",
                            "label",
                            "is_open"
                        ]
                    },
                    "start_date": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Ng\u00e0y thu\u1ea7n, kh\u00f4ng k\u00e8m gi\u1edd: m\u1ed1c d\u1ef1 \u00e1n l\u00e0 ng\u00e0y l\u00e0m vi\u1ec7c, kh\u00f4ng ph\u1ea3i\nm\u1ed9t th\u1eddi \u0111i\u1ec3m \u2014 xem README \"Quy \u01b0\u1edbc d\u1eef li\u1ec7u, th\u1eddi gian & ti\u1ec1n t\u1ec7\"."
                    },
                    "end_date": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "owner": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "properties": {
                            "id": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name",
                            "email"
                        ]
                    },
                    "department": {
                        "type": [
                            "object",
                            "null"
                        ],
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "id",
                            "name"
                        ]
                    },
                    "task_count": {
                        "type": "integer"
                    },
                    "member_count": {
                        "type": "integer"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "code",
                    "description",
                    "status",
                    "start_date",
                    "end_date",
                    "created_at",
                    "updated_at"
                ],
                "title": "ProjectResource"
            },
...
            "StoreCommentRequest": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 20000
                    },
                    "parent_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uuid",
                        "description": "Nh\u1eadn uuid ch\u1ee9 kh\u00f4ng nh\u1eadn id tu\u1ea7n t\u1ef1. B\u00ecnh lu\u1eadn cha ph\u1ea3i thu\u1ed9c\n\u0111\u00fang task n\u00e0y \u2014 ki\u1ec3m \u1edf controller v\u00ec c\u1ea7n bi\u1ebft task n\u00e0o."
                    }
                },
                "required": [
                    "body"
                ],
                "title": "StoreCommentRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}